separate PPND logic out of pipelinecontroller file
